[APPC-3399] removed first screen and bottom sheet from the onboarding
