+++ app/src/main/java/com/asfoundation/wallet/onboarding/OnboardingContent.kt
@@ -1,5 +1,5 @@
 package com.asfoundation.wallet.onboarding
 
 enum class OnboardingContent {
-  EMPTY, WELCOME, VALUES
+  EMPTY, VALUES
 }
