+++ app/src/main/java/com/asfoundation/wallet/onboarding/OnboardingFragment.kt
@@ -1,20 +1,28 @@
 package com.asfoundation.wallet.onboarding
 
-import android.content.res.Configuration
+import android.graphics.Typeface
+import android.net.Uri
 import android.os.Bundle
+import android.text.SpannableString
+import android.text.Spanned
+import android.text.TextPaint
+import android.text.method.LinkMovementMethod
+import android.text.style.ClickableSpan
+import android.text.style.StyleSpan
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.activity.OnBackPressedCallback
 import androidx.annotation.Nullable
+import androidx.core.content.res.ResourcesCompat
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import by.kirich1409.viewbindingdelegate.viewBinding
+import com.asf.wallet.BuildConfig
 import com.asf.wallet.R
 import com.asf.wallet.databinding.FragmentOnboardingBinding
 import com.asfoundation.wallet.base.SingleStateFragment
 import com.asfoundation.wallet.my_wallets.create_wallet.CreateWalletDialogFragment
-import com.asfoundation.wallet.onboarding.bottom_sheet.TermsConditionsBottomSheetFragment
 import com.asfoundation.wallet.viewmodel.BasePageViewFragment
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
@@ -29,14 +37,17 @@ class OnboardingFragment : BasePageViewFragment(),
   private val viewModel: OnboardingViewModel by viewModels()
   private val views by viewBinding(FragmentOnboardingBinding::bind)
   private val onBackPressedCallback = object : OnBackPressedCallback(true) {
-    override fun handleOnBackPressed() = viewModel.handleBackButtonClick()
+    override fun handleOnBackPressed() {
+      isEnabled = false
+      activity?.onBackPressed()
+      activity?.finishAffinity()
+    }
   }
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     handleBackPress()
     handleWalletCreationFragmentResult()
-    handleTermsConditionsFragmentResult()
   }
 
   private fun handleBackPress() {
@@ -46,15 +57,6 @@ class OnboardingFragment : BasePageViewFragment(),
   private fun handleWalletCreationFragmentResult() {
     parentFragmentManager.setFragmentResultListener(
       CreateWalletDialogFragment.CREATE_WALLET_DIALOG_COMPLETE,
-      this
-    ) { _, _ ->
-      navigator.navigateToNavBar()
-    }
-  }
-
-  private fun handleTermsConditionsFragmentResult() {
-    parentFragmentManager.setFragmentResultListener(
-      TermsConditionsBottomSheetFragment.TERMS_CONDITIONS_COMPLETE,
       this
     ) { _, _ ->
       navigator.navigateToNavBar()
@@ -71,63 +73,88 @@ class OnboardingFragment : BasePageViewFragment(),
   override fun onViewCreated(view: View, @Nullable savedInstanceState: Bundle?) {
     super.onViewCreated(view, savedInstanceState)
     setClickListeners()
+    setStringWithLinks()
     viewModel.collectStateAndEvents(lifecycle, viewLifecycleOwner.lifecycleScope)
   }
 
   private fun setClickListeners() {
-    views.onboardingWelcomeButtons.onboardingNextButton.setOnClickListener { viewModel.handleNextClick() }
-    views.onboardingWelcomeButtons.onboardingExistentWalletButton.setOnClickListener { viewModel.handleRecoverClick() }
-
-    views.onboardingValuesButtons.onboardingBackButton.setOnClickListener { viewModel.handleBackButtonClick() }
-    views.onboardingValuesButtons.onboardingGetStartedButton.setOnClickListener { viewModel.handleGetStartedClick() }
+    views.onboardingButtons.onboardingNextButton.setOnClickListener { viewModel.handleLaunchWalletClick() }
+    views.onboardingButtons.onboardingExistentWalletButton.setOnClickListener { viewModel.handleRecoverClick() }
   }
 
   override fun onStateChanged(state: OnboardingState) {
     when (state.pageContent) {
       OnboardingContent.EMPTY -> hideContent()
-      OnboardingContent.WELCOME -> showWelcomeScreen()
       OnboardingContent.VALUES -> showValuesScreen()
     }
   }
 
   override fun onSideEffect(sideEffect: OnboardingSideEffect) {
     when (sideEffect) {
-      OnboardingSideEffect.NavigateToLegalsConsent -> navigator.navigateToTermsBottomSheet()
       OnboardingSideEffect.NavigateToRecoverWallet -> navigator.navigateToRecover()
       OnboardingSideEffect.NavigateToWalletCreationAnimation -> navigator.navigateToCreateWalletDialog()
-      OnboardingSideEffect.NavigateToExit -> {
-        onBackPressedCallback.isEnabled = false
-        activity?.onBackPressed()
-        activity?.finishAffinity()
-      }
+      OnboardingSideEffect.NavigateToFinish -> navigator.navigateToNavBar()
+      is OnboardingSideEffect.NavigateToLink -> navigator.navigateToBrowser(sideEffect.uri)
     }
-  }
-
-  private fun showWelcomeScreen() {
-    views.onboardingValuePropositions.root.visibility = View.GONE
-    views.onboardingValuesButtons.root.visibility = View.GONE
-    views.onboardingWalletIcon.visibility = View.VISIBLE
-    views.onboardingWelcomeMessage.root.visibility = View.VISIBLE
-    views.onboardingWelcomeButtons.root.visibility = View.VISIBLE
   }
 
   private fun showValuesScreen() {
-    if (resources.configuration.orientation == Configuration.ORIENTATION_PORTRAIT) {
-      views.onboardingWalletIcon.visibility = View.VISIBLE
-    } else {
-      views.onboardingWalletIcon.visibility = View.GONE
-    }
-    views.onboardingWelcomeMessage.root.visibility = View.GONE
-    views.onboardingWelcomeButtons.root.visibility = View.GONE
+    views.onboardingWalletIcon?.visibility = View.VISIBLE
     views.onboardingValuePropositions.root.visibility = View.VISIBLE
-    views.onboardingValuesButtons.root.visibility = View.VISIBLE
+    views.onboardingButtons.root.visibility = View.VISIBLE
+    views.onboardingTermsConditions.root.visibility = View.VISIBLE
   }
 
   private fun hideContent() {
+    views.onboardingWalletIcon?.visibility = View.GONE
     views.onboardingValuePropositions.root.visibility = View.GONE
-    views.onboardingValuesButtons.root.visibility = View.GONE
-    views.onboardingWalletIcon.visibility = View.GONE
-    views.onboardingWelcomeMessage.root.visibility = View.GONE
-    views.onboardingWelcomeButtons.root.visibility = View.GONE
+    views.onboardingButtons.root.visibility = View.GONE
+    views.onboardingTermsConditions.root.visibility = View.GONE
+  }
+
+  private fun setStringWithLinks() {
+    val termsConditions = resources.getString(R.string.terms_and_conditions)
+    val privacyPolicy = resources.getString(R.string.privacy_policy)
+    val termsPolicyTickBox =
+      resources.getString(
+        R.string.terms_and_conditions_tickbox, termsConditions,
+        privacyPolicy
+      )
+
+    val spannableString = SpannableString(termsPolicyTickBox)
+    setLinkToString(spannableString, termsConditions, Uri.parse(BuildConfig.TERMS_CONDITIONS_URL))
+    setLinkToString(spannableString, privacyPolicy, Uri.parse(BuildConfig.PRIVACY_POLICY_URL))
+
+    views.onboardingTermsConditions.termsConditionsBody.text = spannableString
+    views.onboardingTermsConditions.termsConditionsBody.isClickable = true
+    views.onboardingTermsConditions.termsConditionsBody.movementMethod =
+      LinkMovementMethod.getInstance()
+  }
+
+  private fun setLinkToString(
+    spannableString: SpannableString, highlightString: String,
+    uri: Uri
+  ) {
+    val clickableSpan = object : ClickableSpan() {
+      override fun onClick(widget: View) {
+        viewModel.handleLinkClick(uri = uri)
+      }
+
+      override fun updateDrawState(ds: TextPaint) {
+        ds.color = ResourcesCompat.getColor(resources, R.color.appc_pink, null)
+        ds.isUnderlineText = true
+      }
+    }
+    val indexHighlightString = spannableString.toString()
+      .indexOf(highlightString)
+    val highlightStringLength = highlightString.length
+    spannableString.setSpan(
+      clickableSpan, indexHighlightString,
+      indexHighlightString + highlightStringLength, Spanned.SPAN_EXCLUSIVE_EXCLUSIVE
+    )
+    spannableString.setSpan(
+      StyleSpan(Typeface.BOLD), indexHighlightString,
+      indexHighlightString + highlightStringLength, Spanned.SPAN_EXCLUSIVE_EXCLUSIVE
+    )
   }
 }
