+++ app/src/main/java/com/asfoundation/wallet/onboarding/OnboardingNavigator.kt
@@ -1,23 +1,23 @@
 package com.asfoundation.wallet.onboarding
 
+import android.content.ActivityNotFoundException
+import android.content.Intent
+import android.net.Uri
+import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
+import com.asf.wallet.R
 import com.asfoundation.wallet.base.Navigator
 import com.asfoundation.wallet.base.navigate
-import com.asfoundation.wallet.onboarding.bottom_sheet.TermsConditionsBottomSheetFragmentDirections
 import javax.inject.Inject
 
 class OnboardingNavigator @Inject constructor(private val fragment: Fragment) : Navigator {
 
-  fun navigateToTermsBottomSheet() {
+  fun navigateToNavBar() {
     navigate(
       fragment.findNavController(),
-      OnboardingFragmentDirections.actionNavigateTermsConditions()
+      OnboardingFragmentDirections.actionNavigateToNavBarFragment()
     )
-  }
-
-  fun navigateToNavBar(){
-    navigate(fragment.findNavController(), OnboardingFragmentDirections.actionNavigateToNavBarFragment())
   }
 
   fun navigateToRecover() {
@@ -34,5 +34,18 @@ class OnboardingNavigator @Inject constructor(private val fragment: Fragment) : 
         needsWalletCreation = true
       )
     )
+  }
+
+  fun navigateToBrowser(uri: Uri) {
+    try {
+      val launchBrowser = Intent(Intent.ACTION_VIEW, uri)
+      launchBrowser.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+      fragment.requireContext()
+        .startActivity(launchBrowser)
+    } catch (exception: ActivityNotFoundException) {
+      exception.printStackTrace()
+      Toast.makeText(fragment.requireContext(), R.string.unknown_error, Toast.LENGTH_SHORT)
+        .show()
+    }
   }
 }
