+++ app/src/main/java/com/asfoundation/wallet/onboarding/OnboardingViewModel.kt
@@ -1,5 +1,6 @@
 package com.asfoundation.wallet.onboarding
 
+import android.net.Uri
 import androidx.lifecycle.viewModelScope
 import com.asfoundation.wallet.app_start.AppStartUseCase
 import com.asfoundation.wallet.app_start.StartMode
@@ -8,16 +9,17 @@ import com.asfoundation.wallet.base.RxSchedulers
 import com.asfoundation.wallet.base.SideEffect
 import com.asfoundation.wallet.base.ViewState
 import com.asfoundation.wallet.onboarding.use_cases.HasWalletUseCase
+import com.asfoundation.wallet.onboarding.use_cases.SetOnboardingCompletedUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 sealed class OnboardingSideEffect : SideEffect {
+  data class NavigateToLink(val uri: Uri) : OnboardingSideEffect()
   object NavigateToWalletCreationAnimation : OnboardingSideEffect()
   object NavigateToRecoverWallet : OnboardingSideEffect()
-  object NavigateToLegalsConsent : OnboardingSideEffect()
-  object NavigateToExit : OnboardingSideEffect()
+  object NavigateToFinish : OnboardingSideEffect()
 }
 
 data class OnboardingState(val pageContent: OnboardingContent = OnboardingContent.EMPTY) : ViewState
@@ -26,6 +28,7 @@ data class OnboardingState(val pageContent: OnboardingContent = OnboardingConten
 class OnboardingViewModel @Inject constructor(
   private val hasWalletUseCase: HasWalletUseCase,
   private val rxSchedulers: RxSchedulers,
+  private val setOnboardingCompletedUseCase: SetOnboardingCompletedUseCase,
   appStartUseCase: AppStartUseCase
 ) :
   BaseViewModel<OnboardingState, OnboardingSideEffect>(initialState()) {
@@ -36,47 +39,42 @@ class OnboardingViewModel @Inject constructor(
     }
   }
 
-  var rootPage = OnboardingContent.EMPTY
-
   init {
+    handleLaunchMode(appStartUseCase)
+  }
+
+  fun handleLaunchMode(appStartUseCase: AppStartUseCase) {
     viewModelScope.launch {
       val mode = appStartUseCase.startModes.first()
-      hasWalletUseCase()
-        .observeOn(rxSchedulers.main)
-        .doOnSuccess {
-          if (!it) {
-            rootPage = OnboardingContent.WELCOME
-            when (mode) {
-              is StartMode.PendingPurchaseFlow -> sendSideEffect { OnboardingSideEffect.NavigateToWalletCreationAnimation }
-              is StartMode.GPInstall -> sendSideEffect { OnboardingSideEffect.NavigateToWalletCreationAnimation }
-              else -> setState { copy(pageContent = OnboardingContent.WELCOME) }
-            }
+      when (mode) {
+        is StartMode.PendingPurchaseFlow -> sendSideEffect { OnboardingSideEffect.NavigateToWalletCreationAnimation }
+        is StartMode.GPInstall -> sendSideEffect { OnboardingSideEffect.NavigateToWalletCreationAnimation }
+        else -> setState { copy(pageContent = OnboardingContent.VALUES) }
+      }
+    }
+  }
+
+  fun handleLaunchWalletClick() {
+    hasWalletUseCase()
+      .observeOn(rxSchedulers.main)
+      .doOnSuccess {
+        setOnboardingCompletedUseCase()
+        sendSideEffect {
+          if (it) {
+            OnboardingSideEffect.NavigateToFinish
           } else {
-            rootPage = OnboardingContent.VALUES
-            setState { copy(pageContent = OnboardingContent.VALUES) }
+            OnboardingSideEffect.NavigateToWalletCreationAnimation
           }
         }
-        .scopedSubscribe { it.printStackTrace() }
-    }
-  }
-
-  fun handleBackButtonClick() = state.pageContent.run {
-    if (this == rootPage) {
-      sendSideEffect { OnboardingSideEffect.NavigateToExit }
-    } else {
-      setState { copy(pageContent = OnboardingContent.WELCOME) }
-    }
-  }
-
-  fun handleNextClick() {
-    setState { copy(pageContent = OnboardingContent.VALUES) }
-  }
-
-  fun handleGetStartedClick() {
-    sendSideEffect { OnboardingSideEffect.NavigateToLegalsConsent }
+      }
+      .scopedSubscribe { it.printStackTrace() }
   }
 
   fun handleRecoverClick() {
     sendSideEffect { OnboardingSideEffect.NavigateToRecoverWallet }
+  }
+
+  fun handleLinkClick(uri: Uri) {
+    sendSideEffect { OnboardingSideEffect.NavigateToLink(uri) }
   }
 }
