+++ app/src/main/res/drawable/white_app_ic_with_text.xml
@@ -1,0 +1,23 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="79dp"
+    android:height="75.809dp"
+    android:viewportWidth="79"
+    android:viewportHeight="75.809">
+  <group>
+    <clip-path
+        android:pathData="M34.865,67.69a5.171,5.171 0,0 1,1.335 0.158,2.6 2.6,0 0,1 1.006,0.5 2.229,2.229 0,0 1,0.637 0.866,3.259 3.259,0 0,1 0.227,1.286v4.913a0.265,0.265 0,0 1,-0.083 0.192,0.26 0.26,0 0,1 -0.191,0.083h-1.534a0.26,0.26 0,0 1,-0.191 -0.083,0.265 0.265,0 0,1 -0.083,-0.192v-0.22a3.485,3.485 0,0 1,-0.841 0.454,3.04 3.04,0 0,1 -1.061,0.165 3.23,3.23 0,0 1,-0.87 -0.117,1.978 1.978,0 0,1 -0.746,-0.392 1.95,1.95 0,0 1,-0.526 -0.73,2.806 2.806,0 0,1 -0.2,-1.128v-0.125a2,2 0,0 1,0.8 -1.768,4.6 4.6,0 0,1 2.569,-0.558h0.739v-0.5q0,-0.591 -0.273,-0.77a1.368,1.368 0,0 0,-0.753 -0.179q-0.654,0 -1.308,0.028 -0.308,0.013 -0.555,0.033l-0.41,0.036h-0.069a0.276,0.276 0,0 1,-0.178 -0.056,0.229 0.229,0 0,1 -0.069,-0.191v-1.115a0.3,0.3 0,0 1,0.26 -0.3q0.383,-0.1 1.034,-0.192A9.109,9.109 0,0 1,34.865 67.69ZM58.376,66.107a0.254,0.254 0,0 1,0.178 0.061,0.219 0.219,0 0,1 0.069,0.172v1.46h1.123a0.26,0.26 0,0 1,0.191 0.082,0.264 0.264,0 0,1 0.083,0.193v1.115a0.263,0.263 0,0 1,-0.083 0.191,0.261 0.261,0 0,1 -0.191,0.084h-1.123v4a1.254,1.254 0,0 0,0.069 0.529c0.045,0.077 0.174,0.117 0.384,0.117h0.671c0.191,0 0.287,0.078 0.287,0.234v1.073c0,0.156 -0.087,0.247 -0.26,0.275s-0.365,0.06 -0.548,0.082a4.38,4.38 0,0 1,-0.534 0.028,5.029 5.029,0 0,1 -0.993,-0.089 1.494,1.494 0,0 1,-0.719 -0.351,1.649 1.649,0 0,1 -0.438,-0.722 3.916,3.916 0,0 1,-0.15 -1.191v-3.99l-1.1,-0.179a0.487,0.487 0,0 1,-0.2 -0.1,0.234 0.234,0 0,1 -0.089,-0.185v-0.909a0.281,0.281 0,0 1,0.089 -0.2,0.273 0.273,0 0,1 0.2,-0.089h1.095v-1.128a0.223,0.223 0,0 1,0.083 -0.179,0.48 0.48,0 0,1 0.191,-0.1l1.671,-0.289ZM50.694,67.69a3.687,3.687 0,0 1,1.473 0.275,2.961 2.961,0 0,1 1.061,0.756 3.177,3.177 0,0 1,0.637 1.143,4.694 4.694,0 0,1 0.212,1.444v0.8c0,0.221 -0.1,0.33 -0.287,0.33h-4.369v0.027a2.287,2.287 0,0 0,0.069 0.564,1.234 1.234,0 0,0 0.232,0.483 1.186,1.186 0,0 0,0.431 0.33,1.625 1.625,0 0,0 0.678,0.124c0.219,0 0.452,0 0.7,-0.014l0.733,-0.028c0.241,-0.009 0.468,-0.02 0.678,-0.033s0.383,-0.031 0.52,-0.049h0.041c0.174,0 0.261,0.073 0.261,0.22v1.045a0.452,0.452 0,0 1,-0.055 0.248,0.368 0.368,0 0,1 -0.232,0.124 11.582,11.582 0,0 1,-1.288 0.248,11.771 11.771,0 0,1 -1.519,0.082 4.418,4.418 0,0 1,-1.233 -0.179,3.024 3.024,0 0,1 -1.109,-0.592 2.966,2.966 0,0 1,-0.794 -1.093,4.125 4.125,0 0,1 -0.3,-1.672v-0.963a4.593,4.593 0,0 1,0.232 -1.5,3.108 3.108,0 0,1 0.679,-1.142 3.01,3.01 0,0 1,1.088 -0.722A3.952,3.952 0,0 1,50.694 67.69ZM21.5,65.461c0.128,0 0.2,0.086 0.219,0.261l0.781,7.376c0.01,0.12 0.036,0.179 0.083,0.179s0.069,-0.059 0.1,-0.178l0.986,-3.786a0.816,0.816 0,0 1,0.343 -0.529,1.022 1.022,0 0,1 0.506,-0.131h1.356a1.026,1.026 0,0 1,0.506 0.131,0.818 0.818,0 0,1 0.343,0.529l0.985,3.785c0.028,0.12 0.06,0.179 0.1,0.179s0.073,-0.059 0.082,-0.178l0.781,-7.377c0.019,-0.175 0.092,-0.261 0.22,-0.261h1.862a0.338,0.338 0,0 1,0.15 0.033,0.149 0.149,0 0,1 0.069,0.145v0.028l-1.274,9.013a1.146,1.146 0,0 1,-0.4 0.827,1.483 1.483,0 0,1 -0.862,0.22h-0.781a1.466,1.466 0,0 1,-0.787 -0.207,1.15 1.15,0 0,1 -0.472,-0.729l-1.082,-4.116a0.2,0.2 0,0 0,-0.048 -0.1,0.148 0.148,0 0,0 -0.075,-0.015 0.108,0.108 0,0 0,-0.069 0.015,0.344 0.344,0 0,0 -0.041,0.1l-1.082,4.116a1.15,1.15 0,0 1,-0.472 0.729,1.468 1.468,0 0,1 -0.787,0.207h-0.781a1.483,1.483 0,0 1,-0.862 -0.22,1.147 1.147,0 0,1 -0.4,-0.827l-1.273,-9.013v-0.026a0.143,0.143 0,0 1,0.069 -0.144,0.327 0.327,0 0,1 0.15,-0.034ZM41.67,65.13a0.24,0.24 0,0 1,0.186 0.083,0.274 0.274,0 0,1 0.075,0.192v9.99a0.3,0.3 0,0 1,-0.075 0.2,0.232 0.232,0 0,1 -0.185,0.089L40,75.684a0.274,0.274 0,0 1,-0.2 -0.089,0.278 0.278,0 0,1 -0.088,-0.2v-9.989a0.256,0.256 0,0 1,0.088 -0.193,0.286 0.286,0 0,1 0.2,-0.083ZM45.532,65.13a0.243,0.243 0,0 1,0.186 0.083,0.279 0.279,0 0,1 0.075,0.192v9.99a0.3,0.3 0,0 1,-0.075 0.2,0.234 0.234,0 0,1 -0.186,0.089h-1.67a0.274,0.274 0,0 1,-0.2 -0.089,0.279 0.279,0 0,1 -0.088,-0.2v-9.989a0.256,0.256 0,0 1,0.088 -0.193,0.286 0.286,0 0,1 0.2,-0.083ZM35.893,72.313h-0.781a2.34,2.34 0,0 0,-0.589 0.062,0.877 0.877,0 0,0 -0.369,0.185 0.658,0.658 0,0 0,-0.192 0.317,1.616 1.616,0 0,0 -0.055,0.44v0.125c0,0.294 0.072,0.481 0.213,0.565a1.334,1.334 0,0 0,0.65 0.123,1.963 1.963,0 0,0 0.6,-0.089 3.709,3.709 0,0 0,0.527 -0.214ZM50.681,69.479a1.167,1.167 0,0 0,-0.918 0.385,1.673 1.673,0 0,0 -0.343,1.142v0.056h2.493v-0.056a1.714,1.714 0,0 0,-0.329 -1.149A1.133,1.133 0,0 0,50.681 69.479ZM67.431,11.627a39.832,39.832 0,0 1,0 56.137,1.575 1.575,0 0,1 -1.544,0.449 1.584,1.584 0,0 1,-1.137 -1.143,1.593 1.593,0 0,1 0.447,-1.551 36.648,36.648 0,0 0,0 -51.646,36.217 36.217,0 0,0 -51.394,0 36.647,36.647 0,0 0,0 51.646,1.593 1.593,0 0,1 0.435,1.546 1.583,1.583 0,0 1,-1.13 1.135,1.575 1.575,0 0,1 -1.538,-0.436 39.832,39.832 0,0 1,0 -56.137A39.361,39.361 0,0 1,67.431 11.627ZM39.821,15.349a5.936,5.936 0,0 1,3.975 1.316,8.513 8.513,0 0,1 2.53,3.516h0l3.558,9.558h6.546a1.956,1.956 0,0 1,1.9 1.96,1.955 1.955,0 0,1 -1.9,1.959h-5.086l0.778,1.985h4.367a1.992,1.992 0,0 1,1.986 1.961,1.93 1.93,0 0,1 -0.548,1.384 1.911,1.911 0,0 1,-1.367 0.576h0l-2.979,0h0l4.721,12.785a13.432,13.432 0,0 1,0.542 2.225,3.86 3.86,0 0,1 -0.542,2.076 7.192,7.192 0,0 1,-1.472 1.44,4.109 4.109,0 0,1 -2.115,0.531 4.328,4.328 0,0 1,-4.389 -2.984h0l-1.858,-5.589L31.328,50.048l-1.857,5.462a4.311,4.311 0,0 1,-4.39 3.136,4.838 4.838,0 0,1 -3.2,-1.164 4.148,4.148 0,0 1,-1.341 -2.858,6.64 6.64,0 0,1 0.515,-2.276h0l4.841,-12.783 -2.989,-0.058a1.945,1.945 0,0 1,-1.9 -1.949,1.909 1.909,0 0,1 1.9,-1.914h4.48l0.756,-1.934 -5.009,-0.037a1.961,1.961 0,0 1,-1.943 -1.949,1.921 1.921,0 0,1 0.552,-1.366 1.9,1.9 0,0 1,1.354 -0.567h6.537l3.655,-9.609a8.714,8.714 0,0 1,2.529 -3.516A6.14,6.14 0,0 1,39.821 15.348ZM39.821,24.403 L34.116,41.524h11.512Z"
+        android:fillType="evenOdd"/>
+    <group>
+      <clip-path
+          android:pathData="M-222.638,268.123L288.072,268.123L288.072,-19.152L-222.638,-19.152Z"/>
+      <path
+          android:pathData="M-222.638,-32.718L1959.053,-32.718L1959.053,1066.107L-222.638,1066.107Z"
+          android:strokeAlpha="0"
+          android:fillColor="#fff"
+          android:fillAlpha="0"/>
+      <path
+          android:pathData="M-222.638,-32.718L1959.053,-32.718L1959.053,1066.107L-222.638,1066.107Z"
+          android:fillColor="#fff"/>
+    </group>
+  </group>
+</vector>
