+++ app/src/main/res/layout/fragment_onboarding.xml
@@ -14,7 +14,7 @@
       android:layout_height="0dp"
       android:layout_marginBottom="@dimen/half_large_margin"
       app:layout_constraintTop_toTopOf="parent"
-      app:layout_constraintBottom_toTopOf="@id/onboarding_buttons"
+      app:layout_constraintBottom_toTopOf="@id/onboarding_action"
       >
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -25,26 +25,11 @@
           android:id="@+id/onboarding_wallet_icon"
           android:layout_width="@dimen/onboarding_icon_size"
           android:layout_height="@dimen/onboarding_icon_size"
-          android:layout_marginTop="72dp"
-          android:src="@drawable/ic_wallet_black"
+          android:layout_marginTop="@dimen/onboarding_container_margin_top"
+          android:src="@drawable/white_app_ic_with_text"
           app:layout_constraintEnd_toEndOf="parent"
           app:layout_constraintStart_toStartOf="parent"
           app:layout_constraintTop_toTopOf="parent"
-          app:tint="@color/white"
-          />
-
-      <include
-          android:id="@+id/onboarding_welcome_message"
-          layout="@layout/onboarding_welcome_message"
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:layout_marginHorizontal="@dimen/half_large_margin"
-          android:layout_marginTop="@dimen/onboarding_container_margin_top"
-          android:visibility="visible"
-          app:layout_constraintBottom_toBottomOf="parent"
-          app:layout_constraintEnd_toEndOf="parent"
-          app:layout_constraintStart_toStartOf="parent"
-          app:layout_constraintTop_toBottomOf="@id/onboarding_wallet_icon"
           app:tint="@color/white"
           />
 
@@ -55,7 +40,7 @@
           android:layout_height="wrap_content"
           android:layout_marginHorizontal="@dimen/half_large_margin"
           android:layout_marginTop="@dimen/onboarding_container_margin_top"
-          android:visibility="gone"
+          android:visibility="visible"
           app:layout_constraintEnd_toEndOf="parent"
           app:layout_constraintStart_toStartOf="parent"
           app:layout_constraintTop_toBottomOf="@id/onboarding_wallet_icon"
@@ -64,31 +49,30 @@
     </androidx.constraintlayout.widget.ConstraintLayout>
   </ScrollView>
   <androidx.constraintlayout.widget.ConstraintLayout
-      android:id="@+id/onboarding_buttons"
+      android:id="@+id/onboarding_action"
       android:layout_width="match_parent"
       android:layout_height="wrap_content"
+      android:layout_marginHorizontal="@dimen/half_large_margin"
       app:layout_constraintBottom_toBottomOf="parent"
       app:layout_constraintTop_toBottomOf="@id/onboarding_content_scroll_view"
       >
 
     <include
-        android:id="@+id/onboarding_welcome_buttons"
-        layout="@layout/onboarding_welcome_buttons"
+        android:id="@+id/onboarding_buttons"
+        layout="@layout/onboarding_buttons"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:visibility="visible"
+        />
+
+    <include
+        android:id="@+id/onboarding_terms_conditions"
+        layout="@layout/onboarding_terms_conditions"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginHorizontal="@dimen/half_large_margin"
         android:visibility="visible"
-        app:tint="@color/white"
-        />
-
-    <include
-        android:id="@+id/onboarding_values_buttons"
-        layout="@layout/onboarding_values_buttons"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="@dimen/half_large_margin"
-        android:visibility="gone"
-        app:tint="@color/white"
+        app:layout_constraintTop_toBottomOf="@id/onboarding_buttons"
         />
   </androidx.constraintlayout.widget.ConstraintLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
