+++ app/src/main/res/layout/onboarding_buttons.xml
@@ -1,34 +1,35 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:theme="@style/MaterialAppTheme"
+    tools:background="@color/bottom_nav_top_up_blue"
     >
 
   <com.asfoundation.wallet.ui.common.WalletButtonView
       android:id="@+id/onboarding_next_button"
       android:layout_width="0dp"
       android:layout_height="@dimen/onboarding_button_height"
-      android:layout_marginBottom="30dp"
-      android:layout_marginStart="8dp"
+      android:layout_marginBottom="18dp"
       app:buttonText="@string/action_next"
       app:buttonType="filled"
-      app:layout_constraintBottom_toBottomOf="parent"
+      app:layout_constraintBottom_toTopOf="@+id/onboarding_existent_wallet_button"
       app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toEndOf="@id/onboarding_existent_wallet_button"
+      app:layout_constraintStart_toStartOf="parent"
       />
   <com.asfoundation.wallet.ui.common.WalletButtonView
       android:id="@+id/onboarding_existent_wallet_button"
       android:layout_width="0dp"
       android:layout_height="@dimen/onboarding_button_height"
-      android:layout_marginBottom="30dp"
-      android:layout_marginEnd="8dp"
+      android:layout_marginBottom="@dimen/half_large_margin"      android:layout_marginTop="@dimen/onboarding_button_existent_margin_top"
       app:buttonColor="@color/white"
       app:buttonText="@string/existent_wallet"
       app:buttonType="outlined"
       app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toStartOf="@+id/onboarding_next_button"
+      app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintStart_toStartOf="parent"
       />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
