+++ app/src/main/res/layout/onboarding_terms_conditions.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/terms_conditions_layout"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    tools:background="@color/bottom_nav_top_up_blue"
+    >
+
+  <ImageView
+      android:id="@+id/terms_conditions_checkmark"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:src="@drawable/terms_conditions_checkmark"
+      app:layout_constraintHorizontal_chainStyle="packed"
+      app:layout_constraintTop_toTopOf="@id/terms_conditions_body"
+      app:layout_constraintStart_toStartOf="parent"
+      app:layout_constraintEnd_toStartOf="@id/terms_conditions_body"
+
+      />
+  <TextView
+      android:id="@+id/terms_conditions_body"
+      android:layout_width="0dp"
+      android:layout_height="wrap_content"
+      android:layout_marginStart="@dimen/normal_margin"
+      android:layout_marginBottom="@dimen/half_large_margin"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintStart_toEndOf="@id/terms_conditions_checkmark"
+      app:layout_constraintTop_toTopOf="parent"
+      app:layout_constraintBottom_toBottomOf="parent"
+      tools:text="I agree to AppCoins Wallet’s Terms and Conditions and Privacy Policy"
+      style="@style/OnboardingValuesBodyTextStyle"
+      />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
