--- app/src/main/res/layout/onboarding_top_app_fragment.xml
@@ -1,102 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/onboarding_iap_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/blue_dark_transparent_90"
-    >
-
-  <ImageView
-      android:id="@+id/toolbar_layout_logo"
-      android:layout_width="67dp"
-      android:layout_height="24dp"
-      android:layout_margin="16dp"
-      android:src="@drawable/ic_app_logo"
-      app:layout_constraintLeft_toLeftOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      />
-
-  <ScrollView
-      android:layout_width="match_parent"
-      android:layout_height="match_parent"
-      >
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/onboarding_iap_content"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        >
-
-      <ImageView
-          android:id="@+id/onboarding_iap_game_icon"
-          android:layout_width="90dp"
-          android:layout_height="90dp"
-          android:layout_marginTop="190dp"
-          app:layout_constraintEnd_toEndOf="parent"
-          app:layout_constraintStart_toStartOf="parent"
-          app:layout_constraintTop_toTopOf="parent"
-          />
-
-      <TextView
-          android:id="@+id/onboarding_iap_title"
-          style="@style/TextAppearance.AppCompat.Body1.Roboto.Medium"
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:layout_marginHorizontal="52dp"
-          android:layout_marginTop="@dimen/onboarding_title_margin_top"
-          android:gravity="center_horizontal"
-          android:letterSpacing="0.01"
-          android:text="@string/onboarding_1st_title"
-          android:textColor="@color/white"
-          android:textSize="22sp"
-          app:layout_constraintEnd_toEndOf="parent"
-          app:layout_constraintStart_toStartOf="parent"
-          app:layout_constraintTop_toBottomOf="@id/onboarding_iap_game_icon"
-          />
-      <TextView
-          android:id="@+id/onboarding_iap_description"
-          style="@style/TextAppearance.AppCompat.Body1.Roboto.Regular"
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:layout_marginHorizontal="@dimen/large_margin"
-          android:layout_marginTop="@dimen/onboarding_subtitle_margin_top"
-          android:gravity="center_horizontal"
-          android:text="@string/onboarding_1st_back_to_unknown_game_body"
-          android:textColor="@color/white"
-          android:textSize="16sp"
-          app:layout_constraintEnd_toEndOf="parent"
-          app:layout_constraintStart_toStartOf="parent"
-          app:layout_constraintTop_toBottomOf="@id/onboarding_iap_title"
-          />
-
-      <com.asfoundation.wallet.ui.common.WalletButtonView
-          android:id="@+id/onboarding_iap_back_to_game_button"
-          android:layout_width="0dp"
-          android:layout_height="@dimen/onboarding_button_height"
-          android:layout_marginHorizontal="70dp"
-          android:layout_marginTop="@dimen/half_large_padding"
-          app:buttonText="@string/onboarding_1st_back_to_unknown_game_button"
-          app:buttonType="filled"
-          app:layout_constraintEnd_toEndOf="parent"
-          app:layout_constraintStart_toStartOf="parent"
-          app:layout_constraintTop_toBottomOf="@id/onboarding_iap_description"
-          />
-
-      <com.asfoundation.wallet.ui.common.WalletButtonView
-          android:id="@+id/onboarding_explore_wallet_button"
-          android:layout_width="0dp"
-          android:layout_height="@dimen/onboarding_button_height"
-          android:layout_marginHorizontal="70dp"
-          android:layout_marginTop="@dimen/big_padding"
-          app:buttonColor="@color/grey"
-          app:buttonText="@string/onboarding_1st_explore_wallet_alternative_button"
-          app:buttonType="text"
-          app:layout_constraintEnd_toEndOf="parent"
-          app:layout_constraintStart_toStartOf="parent"
-          app:layout_constraintTop_toBottomOf="@id/onboarding_iap_back_to_game_button"
-          />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-  </ScrollView>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
