--- app/src/main/res/layout/onboarding_values_buttons.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:theme="@style/MaterialAppTheme"
-    >
-
-  <com.asfoundation.wallet.ui.common.WalletButtonView
-      android:id="@+id/onboarding_get_started_button"
-      android:layout_width="0dp"
-      android:layout_height="@dimen/onboarding_button_height"
-      android:layout_marginBottom="30dp"
-      android:layout_marginStart="27dp"
-      android:layout_marginTop="@dimen/onboarding_button_existent_margin_top"
-      app:buttonText="@string/get_started"
-      app:buttonType="filled"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toEndOf="@+id/onboarding_back_button"
-      />
-
-  <com.google.android.material.card.MaterialCardView
-      android:id="@+id/onboarding_back_button"
-      android:layout_width="48dp"
-      android:layout_height="@dimen/onboarding_button_height"
-      android:background="@color/transparent"
-      app:cardBackgroundColor="@color/transparent"
-      app:cardElevation="0dp"
-      app:layout_constraintEnd_toStartOf="@+id/onboarding_get_started_button"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="@+id/onboarding_get_started_button"
-      app:rippleColor="@color/white"
-      app:shapeAppearance="@style/Home.PillShape"
-      app:strokeColor="@color/white"
-      app:strokeWidth="1dp"
-      >
-
-    <ImageView
-        android:layout_width="48dp"
-        android:layout_height="48dp"
-        android:layout_gravity="center"
-        android:src="@drawable/ic_back_button"
-        />
-
-  </com.google.android.material.card.MaterialCardView>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
