--- app/src/main/res/layout/onboarding_welcome_message.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    >
-
-  <TextView
-      android:id="@+id/onboarding_welcome_title"
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:text="@string/intro_1_title"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      style="@style/OnboardingWelcomeTitleTextStyle"
-      />
-  <TextView
-      android:id="@+id/onboarding_welcome_subtitle"
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:layout_marginTop="@dimen/onboarding_subtitle_margin_top"
-      android:text="@string/intro_1_body"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toBottomOf="@id/onboarding_welcome_title"
-      style="@style/OnboardingWelcomeSubtitleTextStyle"
-      />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
