--- app/src/main/res/layout/terms_conditions_bottom_sheet.xml
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/terms_conditions_bottom_sheet_layout"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="@drawable/rounded_dialog"
-    android:orientation="vertical"
-    android:paddingTop="@dimen/onboarding_title_margin_top"
-    android:paddingHorizontal="32dp"
-    android:paddingBottom="32dp"
-    android:clipToPadding="false"
-    android:clipChildren="false"
-    >
-
-  <TextView
-      android:id="@+id/terms_conditions_body"
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      style="@style/BottomSheetEtherTransactionSubtitle"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      tools:text="I agree to AppCoins Wallet’s Terms and Conditions and Privacy Policy"
-      />
-
-  <com.asfoundation.wallet.ui.common.WalletButtonView
-      android:id="@+id/terms_conditions_confirm_button"
-      android:layout_width="wrap_content"
-      android:layout_height="@dimen/large_button_height"
-      app:buttonText="@string/confirm_button"
-      app:buttonType="filled"
-      android:layout_marginTop="@dimen/half_large_margin"
-      android:layout_marginEnd="@dimen/big_margin"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintTop_toBottomOf="@+id/terms_conditions_body"
-      />
-
-  <com.asfoundation.wallet.ui.common.WalletButtonView
-      android:id="@+id/terms_conditions_decline_button"
-      android:layout_width="wrap_content"
-      android:layout_height="@dimen/large_button_height"
-      android:layout_marginEnd="@dimen/big_margin"
-      app:buttonText="@string/decline_button"
-      app:buttonType="text"
-      app:buttonColor="@color/grey_blue_medium"
-      app:layout_constraintBottom_toBottomOf="@+id/terms_conditions_confirm_button"
-      app:layout_constraintEnd_toStartOf="@+id/terms_conditions_confirm_button"
-      app:layout_constraintTop_toTopOf="@+id/terms_conditions_confirm_button"
-      />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
